requirements.txt
django-auth-ldap==1.6.1
django-autocomplete-light==3.3.0rc1
django-bootstrap-datepicker==1.2.3
django-bootstrap-form==3.2.1
django-crispy-forms==1.7.2
django-crudbuilder==0.2.7
django-debug-toolbar==1.9.1
django-extra-views==0.12.0
django-filter==2.0.0
django-scheduler==0.8.8
django-select2-rocks==0.8.0
django-tables2==2.0.2
django-tables2-column-shifter==0.5.2
django-widget-tweaks==1.4.3

